+++ lib/domain/repository/settings.dart
@@ -44,4 +44,7 @@ abstract class AbstractSettingsRepository {
 
   /// Sets the [ApplicationSettings.locale] value.
   Future<void> setLocale(String locale);
+
+  /// Sets the [ApplicationSettings.showIntroduction] value.
+  Future<void> setShowIntroduction(bool show);
 }