+++ lib/ui/widget/context_menu/overlay.dart
@@ -17,6 +17,7 @@
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
+import 'package:get/get.dart';
 
 /// Overlay of a context menu.
 ///
@@ -47,7 +48,7 @@ class ContextMenuOverlay extends StatefulWidget {
 /// State of a [ContextMenuOverlay].
 class ContextMenuOverlayState extends State<ContextMenuOverlay> {
   /// Currently opened context menu.
-  Widget? _menu;
+  final Rx<Widget?> _menu = Rx(null);
 
   /// Size of the [ContextMenuOverlay].
   Size? _area;
@@ -69,6 +70,9 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
   /// - `topLeft`, meaning [_menu] is placed in the top left quadrant.
   Alignment get alignment => _alignment;
 
+  /// Returns this [_menu] widget.
+  Rx<Widget?> get menu => _menu;
+
   @override
   Widget build(BuildContext context) {
     return LayoutBuilder(
@@ -77,7 +81,7 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
         // This is classic context menu behavior at the OS level.
         final Size size = constraints.biggest;
         if (size != _area) {
-          _menu = null;
+          _menu.value = null;
           _area = size;
         }
 
@@ -95,7 +99,7 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
             body: Stack(
               children: [
                 widget.child,
-                if (_menu != null) ...[
+                if (_menu.value != null) ...[
                   // Listens for taps outside the [_menu].
                   Listener(
                     behavior: HitTestBehavior.opaque,
@@ -163,7 +167,7 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
                         _alignment.x > 0 ? 0 : -1,
                         _alignment.y > 0 ? 0 : -1,
                       ),
-                      child: IntrinsicWidth(child: _menu),
+                      child: IntrinsicWidth(child: _menu.value),
                     ),
                   ),
                 ]
@@ -178,11 +182,11 @@ class ContextMenuOverlayState extends State<ContextMenuOverlay> {
   /// Sets the current menu to [child] at [position].
   void show(Widget child, Offset position) => setState(() {
         _position = position;
-        _menu = child;
+        _menu.value = child;
       });
 
   /// Hides the current menu if there is one.
-  void hide() => setState(() => _menu = null);
+  void hide() => setState(() => _menu.value = null);
 }
 
 /// [InheritedWidget] of a [_ContextMenuOverlayState] used to implement